Move credentials configurator to executable script
