+++ dev/google-cloud
@@ -18,7 +18,6 @@
 
 require __DIR__ . '/../vendor/autoload.php';
 
-use Google\Cloud\Dev\Credentials\Credentials;
 use Google\Cloud\Dev\DocGenerator\Command\Docs;
 use Google\Cloud\Dev\Release\Command\Release;
 use Google\Cloud\Dev\Split\Command\Split;
@@ -33,7 +32,6 @@ if (!class_exists(Application::class)) {
 }
 
 $app = new Application;
-$app->add(new Credentials(__DIR__));
 $app->add(new Release(__DIR__));
 $app->add(new Docs(__DIR__));
 $app->add(new Split(__DIR__));
